[REF] rename is_multi_variant_line -> is_configurable_with_variant
