+++ sale_configurator_variant/static/src/js/sol_many2one_inherit.esm.js
@@ -1,0 +1,22 @@
+/** @odoo-module **/
+
+import {SaleOrderLineProductField} from "@sale/js/sale_product_field";
+import {patch} from "@web/core/utils/patch";
+
+patch(SaleOrderLineProductField.prototype, {
+    get productName() {
+        const isConfigurableWithVariant =
+            this.props.record.data.is_configurable_with_variant;
+        const productTemplateData = this.props.record.data.product_template_id;
+
+        if (
+            isConfigurableWithVariant &&
+            productTemplateData &&
+            productTemplateData[1]
+        ) {
+            return productTemplateData[1];
+        }
+
+        return super.productName;
+    },
+});
