+++ sale_configurator_variant/views/sale_view.xml
@@ -8,26 +8,26 @@
         />
         <field name="arch" type="xml">
             <xpath expr="//label[@for='product_id']" position="before">
-                <field name="is_multi_variant_line" />
+                <field name="is_configurable_with_variant" />
                 <field
                     name="product_template_id"
-                    invisible="not is_multi_variant_line"
-                    required="is_multi_variant_line"
+                    invisible="not is_configurable_with_variant"
+                    required="is_configurable_with_variant"
                 />
             </xpath>
             <xpath expr="//field[@name='product_id']/.." position="attributes">
-                <attribute name="invisible">is_multi_variant_line</attribute>
+                <attribute name="invisible">is_configurable_with_variant</attribute>
             </xpath>
             <xpath expr="//label[@for='product_id']" position="attributes">
-                <attribute name="invisible">is_multi_variant_line</attribute>
+                <attribute name="invisible">is_configurable_with_variant</attribute>
             </xpath>
 
             <xpath expr="//field[@name='name']" position="after">
-                <label for="variant_ids" invisible="not is_multi_variant_line" />
+                <label for="variant_ids" invisible="not is_configurable_with_variant" />
                 <field
                     name="variant_ids"
                     context="{'default_order_id': order_id}"
-                    invisible="not is_multi_variant_line"
+                    invisible="not is_configurable_with_variant"
                 >
                     <list editable="bottom">
                         <!-- There is some magic here, please look at models/ir_ui_view.py
@@ -75,7 +75,7 @@
                 position="before"
             >
                 <!-- Need to be present in view for JS patch -->
-                <field name="is_multi_variant_line" column_invisible="1"/>
+                <field name="is_configurable_with_variant" column_invisible="1" />
             </xpath>
         </field>
     </record>
