+++ test_sale_configurator_option_variant/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_sale_order
